added st_sscan subcommand
