+++ limix/_cli/qtl.py
@@ -1,0 +1,66 @@
+import click
+from .st_sscan import st_sscan
+
+
+# @click.command(cls=OrderedCommand)
+@click.group()
+@click.pass_context
+# @click.option(
+#     "--model",
+#     help=("Specify the statistical model to perform the scan."),
+#     default="single-trait",
+#     type=click.Choice(["single-trait", "struct-lmm"]),
+# )
+# @click.option(
+#     "--lik",
+#     help=(
+#         "Specify the type of likelihood that will described"
+#         " the residual error distribution."
+#     ),
+#     default="normal",
+# )
+def qtl(ctx):
+    """Perform genome-wide association scan.
+
+    This analysis requires minimally the specification of one phenotype
+    (PHENOTYPES_FILE) and genotype data (GENOTYPE_FILE).
+
+    The --filter option allows for selecting a subset of the original dataset for
+    the analysis. For example,
+
+        --filter="genotype: (chrom == '3') & (pos > 100) & (pos < 200)"
+
+    states that only loci of chromosome 3 having a position inside the range (100, 200)
+    will be considered. The --filter option can be used multiple times in the same
+    call. In general, --filter accepts a string of the form
+
+        <DATA-TYPE>: <BOOL-EXPR>
+
+    where <DATA-TYPE> can be phenotype, genotype, or covariate. <BOOL-EXPR> is a boolean
+    expression involving row or column names. Please, consult `pandas.DataFrame.query`
+    function from Pandas package for further information.
+    \f
+
+    Examples
+    --------
+
+    ... doctest::
+
+        # First we perform a quick file inspection. This step is optional but is very
+        # useful to check whether `limix` is able to read them and print out their
+        # metadata.
+        limix show phenotypes.csv
+        limix show genotype.bgen
+        limix show kinship.raw
+
+        # We now perform the analysis, specifying the genotype loci and the phenotype
+        # of interest.
+        limix phenotypes.csv genotype.bgen --kinship-file=kinship.raw \
+            --output-dir=results \
+            --filter="phenotype: col == 'height'" \
+            --filter="genotype: (chrom == '3') & (pos > 100) & (pos < 200)"
+    """
+    pass
+
+
+qtl.add_command(st_sscan)
